move is_opening/is_closing/is_calibrated/is_reversed/get_light_level to base class for compat
